Shankjbs571/requirements.txt
django
pypdf #to process .pdf files
python-docx #to process .docx files
# textract
pandas